+++ admin/views/memberships-page.php
@@ -1,44 +1,77 @@
 <?php
+/**
+ * Fapi
+ *
+ * @package   Fapi membership
+ * @author    Vladislav Musílek
+ * @license   GPL-2.0+
+ * @link      http://musilda.com
+ * @copyright 2020 Musilda.com
+ *
+ */
+
 declare(strict_types = 1);
 
 $memberships = Fapi_Memberships::get_instance();
 if ( ! empty( $_POST['save'] ) ) {
-	if ( ! empty( $_POST['membership_name'] ) ) {
-		$data = $memberships->get_memberships();
 
-		if ( empty( $data ) ) {
-			$key = 1;
-		} else {
-			$key = count( $data ) + 1;
+	$nonce = sanitize_text_field( wp_unslash( $_POST['fapi_admin_form'] ) );
+	if ( isset( $nonce ) && wp_verify_nonce( $nonce, 'fapi-admin-form' ) ) {
+
+		if ( ! empty( $_POST['membership_name'] ) ) {
+			$data = $memberships->get_memberships();
+
+			if ( empty( $data ) ) {
+				$key = 1;
+			} else {
+				$key = count( $data ) + 1;
+			}
+
+			if ( ! empty( $_POST['membership_name'] ) ) {
+				$membership_name = sanitize_text_field( wp_unslash( $_POST['membership_name'] ) );
+			}
+			if ( ! empty( $_POST['membership_email'] ) ) {
+				$membership_email = sanitize_text_field( wp_unslash( $_POST['membership_email'] ) );
+			}
+			if ( ! empty( $_POST['membership_redirect'] ) ) {
+				$membership_redirect = sanitize_text_field( wp_unslash( $_POST['membership_redirect'] ) );
+			}
+			if ( ! empty( $_POST['membership_login_redirect'] ) ) {
+				$membership_login_redirect = sanitize_text_field( wp_unslash( $_POST['membership_login_redirect'] ) );
+			}
+
+			$data[ $key ] = array(
+				'name'           => $membership_name,
+				'note'           => $membership_note,
+				'email'          => $membership_email,
+				'redirect'       => $membership_redirect,
+				'login_redirect' => $membership_login_redirect,
+			);
+
+			update_option( 'fapi_memberships', $data );
 		}
-
-
-		$data[ $key ] = array(
-			'name'           => sanitize_text_field( $_POST['membership_name'] ),
-			'note'           => sanitize_text_field( $_POST['membership_note'] ),
-			'email'          => sanitize_text_field( $_POST['membership_email'] ),
-			'redirect'       => sanitize_text_field( $_POST['membership_redirect'] ),
-			'login_redirect' => sanitize_text_field( $_POST['membership_login_redirect'] ),
-		);
-
-		update_option( 'fapi_memberships', serialize( $data ) );
 	}
 }
 
 
 if ( ! empty( $_GET['delete'] ) ) {
 
-	$data = $memberships->get_memberships();
+	$nonce = sanitize_text_field( wp_unslash( $_GET['nonce_delete'] ) );
+	if ( isset( $nonce ) && wp_verify_nonce( $nonce, 'nonce_delete' ) ) {
 
-	if ( ! empty( $data[ $_GET['delete'] ] ) ) {
-		unset( $data[ $_GET['delete'] ] );
-		if ( ! empty( $data ) ) {
-			update_option( 'fapi_memberships', serialize( $data ) );
-		} else {
-			delete_option( 'fapi_memberships' );
+		$data = $memberships->get_memberships();
+		
+		if ( ! empty( $data[ $_GET['delete'] ] ) ) {
+			unset( $data[ $_GET['delete'] ] );
+			if ( ! empty( $data ) ) {
+				update_option( 'fapi_memberships', $data );
+			} else {
+				delete_option( 'fapi_memberships' );
+			}
 		}
+		wp_safe_redirect( admin_url() . 'admin.php?page=fapi-memebership' );
+
 	}
-	wp_redirect( admin_url() . 'admin.php?page=fapi-memebership' );
 
 }
 
@@ -69,14 +102,13 @@ $emails = new WP_Query( $args );
 
 	</div>
 	<div class="clear"></div>
-	
 	<div class="t-col-12">
 		<div class="toret-box box-info">
 			<div class="box-header">
 				<h3 class="box-title"><?php esc_attr_e( 'Add new memberships sections', 'fapi-membership' ); ?></h3>
 			</div>
 			<div class="box-body">
-				<form method="post" action="<?php echo admin_url() . 'admin.php?page=fapi-memebership'; ?>">
+				<form method="post" action="<?php echo esc_url( admin_url() ) . 'admin.php?page=fapi-memebership'; ?>">
 				<table class="table-bordered">
 					<tr>
 						<th><?php esc_attr_e( 'Name', 'fapi-membership' ); ?></th>
@@ -95,7 +127,7 @@ $emails = new WP_Query( $args );
 							<?php
 							if ( ! empty( $emails->posts ) ) {
 								foreach ( $emails->posts as $email ) {
-									echo '<option value="' . $email->ID . '">' . $email->post_title . '</option>';
+									echo '<option value="' . esc_attr( $email->ID ) . '">' . esc_attr( $email->post_title ) . '</option>';
 								}
 							}
 							?>
@@ -106,6 +138,7 @@ $emails = new WP_Query( $args );
 						<td class="td_center"><input class="btn btn-success" type="submit" name="save" value="<?php esc_attr_e( 'Create', 'fapi-membership' ); ?>" /></td>
 					</tr>
 				</table>
+				<?php wp_nonce_field( 'fapi-admin-form', 'fapi_admin_form' ); ?>
 				</form>
 			</div>
 			<div class="clear"></div>
@@ -114,7 +147,4 @@ $emails = new WP_Query( $args );
 
 	</div>
 	<div class="clear"></div>
-
-	
-
 </div>
